requirements.txt
# 수집
feedparser>=6.0.0
requests>=2.31.0
PyYAML>=6.0

# 처리
openai>=1.0.0

# 유틸리티
python-dateutil>=2.8.0
python-dotenv>=1.0.0
